fix: segment out timeline menu features
