+++ src/ui/timeline/timeline-menu/boil-control/boil-control.component.css
@@ -1,0 +1,6 @@
+.boil-icon path {
+  fill: none;
+  stroke: currentColor;
+  stroke-width: 1.6;
+  stroke-linecap: round;
+}
