+++ src/ui/timeline/timeline-menu/boil-control/boil-control.component.html
@@ -1,0 +1,62 @@
+<div class="frame-control boil-control">
+  <button
+    class="frame-control-btn boil-btn"
+    [class.active]="currentBoilEnabled()"
+    [attr.aria-label]="
+      currentBoilEnabled() ? 'Disable line boiling for this frame' : 'Enable line boiling for this frame'
+    "
+    [attr.title]="
+      currentBoilEnabled() ? 'Disable line boiling for this frame' : 'Enable line boiling for this frame'
+    "
+    [attr.aria-pressed]="currentBoilEnabled()"
+    (pointerdown)="onPointerDown($event)"
+    (pointerup)="onPointerUp()"
+    (pointercancel)="onPointerCancel()"
+    (pointerleave)="onPointerCancel()"
+    (contextmenu)="onContextMenu($event)"
+  >
+    <svg viewBox="0 0 24 24" class="frame-control-icon boil-icon" aria-hidden="true" focusable="false">
+      <path d="M3 9c2-2.4 4-2.4 6 0s4 2.4 6 0 4-2.4 6 0"></path>
+      <path d="M3 15c2-2.4 4-2.4 6 0s4 2.4 6 0 4-2.4 6 0"></path>
+    </svg>
+    @if (currentBoilEnabled()) {
+    <span class="submenu-indicator" aria-hidden="true"></span>
+    }
+  </button>
+  @if (submenuOpen()) {
+  <div class="submenu" role="group" aria-label="Line boiling parameters">
+    <div class="submenu-title">Line Boiling</div>
+    <label class="submenu-field">
+      <span>Variations</span>
+      <input
+        type="number"
+        min="2"
+        max="12"
+        [value]="currentBoilParams().variations"
+        (change)="setBoilParam('variations', $any($event.target).value)"
+      />
+    </label>
+    <label class="submenu-field">
+      <span>Hold Frames</span>
+      <input
+        type="number"
+        min="1"
+        max="24"
+        [value]="currentBoilParams().holdFrames"
+        (change)="setBoilParam('holdFrames', $any($event.target).value)"
+      />
+    </label>
+    <label class="submenu-field">
+      <span>Jitter (px)</span>
+      <input
+        type="number"
+        min="0"
+        max="20"
+        step="0.5"
+        [value]="currentBoilParams().amount"
+        (change)="setBoilParam('amount', $any($event.target).value)"
+      />
+    </label>
+  </div>
+  }
+</div>
