+++ src/ui/timeline/timeline-menu/duration-control/duration-control.component.html
@@ -1,0 +1,13 @@
+<label class="duration-control">
+  <span class="control-label">Duration</span>
+  <div class="duration-input-wrap">
+    <input
+      type="number"
+      min="0.1"
+      step="0.1"
+      [value]="currentDuration()"
+      (input)="updateDuration($any($event.target).value)"
+    />
+    <span class="control-unit">s</span>
+  </div>
+</label>
