+++ src/ui/timeline/timeline-menu/frame-control.css
@@ -1,0 +1,113 @@
+/* Shared styles for the frame-toggle controls (boil + onion skin) and their submenus. */
+
+.frame-control {
+  position: relative;
+  display: inline-flex;
+}
+
+.frame-control-btn {
+  position: relative;
+  width: 34px;
+  height: 34px;
+  padding: 0;
+  background-color: transparent;
+  border: 1px solid var(--border-default);
+  border-radius: 6px;
+  cursor: pointer;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  transition:
+    background-color 0.15s,
+    border-color 0.15s;
+  color: var(--text-default);
+}
+
+.frame-control-btn:hover {
+  background-color: var(--bg-hover);
+}
+
+.frame-control-btn.active {
+  background-color: var(--bg-active);
+  border-color: var(--border-active);
+  color: var(--text-on-accent);
+}
+
+.frame-control-icon {
+  width: 21px;
+  height: 21px;
+  display: block;
+}
+
+.submenu-indicator {
+  position: absolute;
+  right: 3px;
+  bottom: 3px;
+  width: 0;
+  height: 0;
+  border-style: solid;
+  border-width: 0 0 6px 6px;
+  border-color: transparent transparent
+    color-mix(in srgb, var(--text-muted) 78%, var(--bg-panel)) transparent;
+  pointer-events: none;
+}
+
+.frame-control-btn:hover .submenu-indicator {
+  border-bottom-color: color-mix(in srgb, var(--text-default) 88%, var(--accent-border));
+}
+
+.frame-control-btn.active .submenu-indicator {
+  border-bottom-color: color-mix(in srgb, var(--text-on-accent) 80%, var(--accent-light));
+}
+
+.submenu {
+  position: absolute;
+  top: calc(100% + 6px);
+  left: 0;
+  z-index: 1000;
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+  min-width: 168px;
+  padding: 10px;
+  background-color: var(--bg-panel);
+  border: 1px solid var(--border-default);
+  border-radius: 8px;
+  box-shadow: 0 8px 24px var(--shadow-sm);
+}
+
+.submenu-title {
+  font-size: 11px;
+  font-weight: 600;
+  text-transform: uppercase;
+  letter-spacing: 0.04em;
+  color: var(--text-muted, var(--text-default));
+  margin-bottom: 2px;
+}
+
+.submenu-field {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: 8px;
+  font-size: 12px;
+  color: var(--text-default);
+}
+
+.submenu-field input {
+  width: 64px;
+  padding: 3px 6px;
+  font-size: 12px;
+  text-align: right;
+  background-color: var(--bg-body);
+  color: var(--text-default);
+  border: 1px solid var(--border-default);
+  border-radius: 4px;
+}
+
+.submenu-field-color input[type='color'] {
+  width: 40px;
+  height: 24px;
+  padding: 1px;
+  cursor: pointer;
+}
