+++ src/ui/timeline/timeline-menu/playback-controls/playback-controls.component.html
@@ -1,0 +1,41 @@
+<div class="playback-controls-group">
+  <button
+    class="back-to-start-btn"
+    (click)="goToStart()"
+    aria-label="Go to start"
+  >
+    <i class="fa-solid fa-step-backward" aria-hidden="true"></i>
+  </button>
+  <button
+    class="back-one-board-btn"
+    (click)="backOneBoard()"
+    aria-label="Back one board"
+  >
+    <i class="fa-solid fa-backward" aria-hidden="true"></i>
+  </button>
+  <button
+    class="play-btn"
+    [class.active]="isPlaying()"
+    (click)="togglePlayback()"
+    aria-label="Play/Pause"
+  >
+    <div class="play-icon" [class.playing]="isPlaying()">
+      <span class="bar bar-left"></span>
+      <span class="bar bar-right"></span>
+    </div>
+  </button>
+  <button
+    class="forward-one-board-btn"
+    (click)="forwardOneBoard()"
+    aria-label="Forward one board"
+  >
+    <i class="fa-solid fa-forward" aria-hidden="true"></i>
+  </button>
+  <button
+    class="forward-to-end-btn"
+    (click)="goToEnd()"
+    aria-label="Go to end"
+  >
+    <i class="fa-solid fa-step-forward" aria-hidden="true"></i>
+  </button>
+</div>
